+++ internal/model/segment.go
@@ -3,11 +3,11 @@ package model
 import "github.com/hashicorp/terraform-plugin-framework/types"
 
 type Segment struct {
-	ID                string           `json:"id"`
-	Name              string           `json:"name"`
-	Selector          string           `json:"selector"`
-	FallbackToDefault bool             `json:"fallback_to_default"`
-	AutoApply         *AutoApplyConfig `json:"auto_apply,omitempty"`
+	ID                string          `json:"id"`
+	Name              string          `json:"name"`
+	Selector          string          `json:"selector"`
+	FallbackToDefault bool            `json:"fallback_to_default"`
+	AutoApply         AutoApplyConfig `json:"auto_apply"`
 }
 
 func (e Segment) ToTF() SegmentTF {